+++ src/core/features/encryption/algorithms.ts
@@ -1,15 +1,9 @@
-export enum ENCRYPTION_ALGORITHM {
-	AES = 'AES',
-	TWOFISH = 'Twofish',
-}
-
-export const buildAlgorithmSequence = (algorithms: ENCRYPTION_ALGORITHM[]) => {
-	return algorithms.join('-');
-};
+import { formatAlgorithmsToString } from './utils';
+import { ENCRYPTION_ALGORITHM } from '.';
 
 export const ENCRYPTION_ALGORITHMS_LIST = [
 	ENCRYPTION_ALGORITHM.AES,
 	ENCRYPTION_ALGORITHM.TWOFISH,
-	buildAlgorithmSequence([ENCRYPTION_ALGORITHM.AES, ENCRYPTION_ALGORITHM.TWOFISH]),
-	buildAlgorithmSequence([ENCRYPTION_ALGORITHM.TWOFISH, ENCRYPTION_ALGORITHM.AES]),
+	formatAlgorithmsToString([ENCRYPTION_ALGORITHM.AES, ENCRYPTION_ALGORITHM.TWOFISH]),
+	formatAlgorithmsToString([ENCRYPTION_ALGORITHM.TWOFISH, ENCRYPTION_ALGORITHM.AES]),
 ];
