+++ src/core/features/encryption/index.ts
@@ -1,0 +1,4 @@
+export enum ENCRYPTION_ALGORITHM {
+	AES = 'AES',
+	TWOFISH = 'Twofish',
+}
