+++ src/core/features/encryption/workers/Cryptography.worker.ts
@@ -11,7 +11,8 @@ import { PipelineProcessor } from '../../../encryption/processors/PipelineProces
 import { getDerivedKeysManager, getMasterKey } from '../../../encryption/utils/keys';
 import { getRandomBytes } from '../../../encryption/utils/random';
 
-import { ENCRYPTION_ALGORITHM } from '../algorithms';
+import { parseAlgorithms } from '../utils';
+import { ENCRYPTION_ALGORITHM } from '..';
 import { FakeWorkerObject } from '.';
 
 export default FakeWorkerObject;
@@ -21,21 +22,6 @@ console.log('Hello world from worker');
 let encryptionController: EncryptionController | null = null;
 const messenger = new WorkerMessenger(self);
 const requests = new WorkerRPC(messenger);
-
-const parseAlgorithms = (algorithms: string): ENCRYPTION_ALGORITHM[] => {
-	const validAlgorithms = Object.values(ENCRYPTION_ALGORITHM);
-	return algorithms.split('-').map((name) => {
-		const algorithm = validAlgorithms.find((alg) => alg === name);
-		if (!algorithm) {
-			throw new Error(
-				`Unsupported encryption algorithm: "${name}". Supported: ${validAlgorithms.join(
-					', ',
-				)}`,
-			);
-		}
-		return algorithm;
-	});
-};
 
 const workerId = performance.now();
 requests.addHandler('init', async ({ key, salt, algorithm }) => {
